Roles (#8)
Each user has a role. The database is seeded with
the following roles:

Administrator
Property owner
Estate agent
Property agent
Property developer
Other business

The purpose of roles are:
1) To allow the site owner to learn more about
his advertisers
2) To restrict features to certain roles

Roles can be set to be selectable on signup. Only
those that are will be displayed to the
advertiser on the sign up form, and only those
will be valid choices when creating the new user.

The "Other business" role is a miscellaneous role
specifically for directory and banner advertisers
who do not fall under the other roles.