Replace Net::SFTP with expect script
Net::SFTP was reaching 100% CPU at a download rate of 2 MB/sec which is
not acceptable when we need to download TripAdvisor archives of around
1.5 GB.

We drop down to using the native SFTP command.

We need to use an expect script to supply our password as we do not have
password-less access.