Remove exchange rate updating code based on Yahoo! feed
Yahoo! has removed their feed, returning 403 and:

It has come to our attention that this service is being used in
violation of the Yahoo Terms of Service.  As such, the service is being
discontinued.  For all future markets and equities data research, please
refer to finance.yahoo.com.

E.g.:
http://download.finance.yahoo.com/d/quotes.csv?f=sl1d1t1ba&e=.csv&s=EUREUR=X,GBPEUR=X,USDEUR=X,CHFEUR=X,CADEUR=X

https://stackoverflow.com/questions/47064776/has-yahoo-suddenly-today-terminated-its-finance-download-api