Include MCID number for TripAdvisor referral attribution
Roger Gaudreau rgaudreau@tripadvisor.com 26th Jan 2018:

With regard to attribution, did Kelly provide you with an MCID number
yet? If so, you just add that to the TA URL as a query parameter
(m=<MCID value>). If you don’t have that, Chris should be able to
provide you one.  The MCID value is maintained in the session, but Chris
can answer the business question regarding attribution if the traveler
books a different property from the one they originally clicked on your
site.